+++ src/components/input-search/input-search.styles.jsx
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+
+export const SearchInputContainer = styled.div`
+  gap: 1rem;
+  display: flex;
+  margin: 4rem;
+  justify-content: flex-end;
+  align-items: center;
+`;
+
+export const InputBar = styled.input`
+  font-size: 1rem;
+  width: 250px;
+  height: 35px;
+  border: none;
+  border-bottom: 1px solid black;
+`;