+++ src/pages/ProductList.page.jsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export const ProductList = () => {
-  return (
-    <div>
-      <h1>Product List</h1>
-    </div>
-  );
-};