+++ src/pages/product-details/product-details.styles.jsx
@@ -1,0 +1,100 @@
+import styled from "styled-components";
+
+export const ProductDetailsContainer = styled.div`
+  display: grid;
+  grid-template-columns: 45% 45%;
+  margin: 40px 20px;
+  padding: 30px;
+  justify-items: center;
+  align-items: center;
+  border: 1px solid #ccc;
+  border-radius: 5px;
+  box-shadow: 0 0 10px 0 rgba(0, 0, 0, 0.1);
+  @media (max-width: 768px) {
+    grid-template-columns: 100%;
+    grid-template-rows: 25% 80%;
+    border: none;
+    box-shadow: none;
+  }
+`;
+
+export const ProductDetailsImage = styled.img`
+  width: 520px;
+  height: 520px;
+  @media (max-width: 768px) {
+    width: 350px;
+    height: 350px;
+  }
+`;
+
+export const ProductDetailsTable = styled.table`
+  font-size: 1.2rem;
+  margin: 20px 0;
+  width: 100%;
+  border-collapse: collapse;
+  line-height: 1.5;
+  @media (max-width: 768px) {
+    font-size: 1rem;
+  }
+`;
+
+export const ProductDetailsTableBody = styled.tbody`
+  tr {
+    border-bottom: 1px solid #ccc;
+    &:last-child {
+      border-bottom: none;
+    }
+  }
+  td {
+    padding: 10px 0;
+  }
+  td:first-child {
+    font-weight: 600;
+  }
+`;
+
+export const ProductDetailsSelectContainer = styled.div`
+  display: grid;
+  grid-template-columns: 10% 20% 10% 20%;
+  justify-content: space-evenly;
+  align-items: center;
+  margin: 20px 0;
+`;
+
+export const ProductDetailsSelectLabel = styled.label`
+  font-size: 1.2rem;
+  font-weight: bold;
+`;
+
+export const ProductDetailsSelect = styled.select`
+  border-radius: 100%;
+  width: 80px;
+  height: 80px;
+  border: 1px solid #ccc;
+  margin: 0 10px;
+  text-align: center;
+`;
+
+export const ProductDetailsSelectOption = styled.option`
+  font-size: 1rem;
+  background-color: #fff;
+`;
+
+export const PriceContainer = styled.div`
+  display: flex;
+  justify-content: space-evenly;
+  align-items: center;
+  margin: 20px 0;
+  @media (max-width: 768px) {
+    flex-direction: column;
+    align-items: flex-start;
+  }
+`;
+
+export const PriceText = styled.p`
+  font-size: 1.5rem;
+  font-weight: 500;
+  @media (max-width: 768px) {
+    font-size: 1rem;
+  }
+`;