+++ CHANGELOG.md
@@ -14,6 +14,26 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Removed
 
+## [0.3.0] - 2023-02-03
+
+### Added
+
+- Added the product detail page with the mock data
+- Added context of products
+- Added dynamic routing for product detail page
+
+### Changed
+
+- Changed the mock DB for having better product details
+
+## [0.2.0] - 2023-02-03
+
+### Added
+
+- Added the mock db
+- Added product list page with the mock data
+- Added search input component
+
 ## [0.1.1] - 2023-02-02
 
 ### Added